requirements.txt
numpy
pandas
nltk
scikit-learn
scipy
matplotlib
absl-py
pyyaml
pytorch_metric_learning
pytorch-lightning
wandb
openai
tenacity
transformers
matplotlib
ipykernel
seaborn
tqdm
tiktoken
sentence-transformers
gensim
termcolor